pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "EnsembleFFFit"
version = "0.1.0"
description = "Tools to coordinate force field fitting"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Ryan Morelock", email = "rym@ornl.gov" }
]
dependencies = [
    "numpy",  # Add your dependencies here
    "ase",
    "torch-sim-atomistic",
    "pymatgen",
    "pymatgen-analysis-defects",
    "PyYAML",
    "frozendict",
    "mp-api",
    "matminer",
    "pyscal3"
]

[project.scripts]
mp_query = "EnsembleFFFit.generation.materials_project.mp_query_cli:main"
generate_EoS = "EnsembleFFFit.generation.equation_of_state.generate_EoS_cli:main"
generate_defects = "EnsembleFFFit/generation/defects/generate_defects_cli.py:main"
generate_substitutions = "EnsembleFFFit/generation/substitutions/generate_substitutions_cli.py:main"

jaxreaxff_matensemble = "EnsembleFFFit/matensemble/reaxff/jaxreaxff_matensemble_cli.py:main"
lammps_matensemble = "EnsembleFFFit/matensemble/lammps/lammps_matensemble_cli.py:main"

copy_by_pattern = "EnsembleFFFit/utils/copy_by_pattern_cli.py:main"

[tool.setuptools]
packages = ["EnsembleFFFit"]  # Ensure this matches your package directory name

